requirements.txt
alembic==1.0.2
altgraph==0.16.1
amqp==2.3.2
aniso8601==4.0.1
asn1crypto==0.24.0
backcall==0.1.0
beautifulsoup4==4.6.1
billiard==3.5.0.4
bleach==3.0.2
blinker==1.4
celery==4.2.1
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.4.0
cryptography==2.3.1
decorator==4.3.0
defusedxml==0.5.0
Django==2.1.4
django-redis==4.10.0
djangorestframework==3.9.0
dominate==2.3.4
entrypoints==0.2.3
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Cache==0.13.1
Flask-HTTPAuth==3.2.4
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.3.0
Flask-Moment==0.6.0
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-Script-Extras==0.1.4
Flask-SQLAlchemy==2.3.2
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
future==0.16.0
idna==2.7
ipykernel==5.1.0
ipython==7.1.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
itchat==1.2.32
itsdangerous==1.1.0
jedi==0.13.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
kombu==4.2.1
lxml==4.2.5
macholib==1.10
Mako==1.0.7
MarkupSafe==1.0
mistune==0.8.4
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.0
numpy==1.15.4
opencv-python==3.4.4.19
pandocfilters==1.4.2
parso==0.3.1
pbr==5.1.1
pefile==2018.8.8
pickleshare==0.7.5
Pillow==5.2.0
prometheus-client==0.4.2
prompt-toolkit==2.0.6
pycparser==2.18
pygame==1.9.4
Pygments==2.2.0
PyInstaller==3.3.1
PyMySQL==0.9.2
pypng==0.0.18
PyQRCode==1.2.1
python-dateutil==2.7.5
python-editor==1.0.3
pytz==2018.5
pywin32==223
pywinpty==0.5.4
pyzmq==17.1.2
qrcode==6.0
qtconsole==4.4.2
redis==3.0.1
requestes==0.0.1
requests==2.19.1
selenium==3.14.0
Send2Trash==1.5.0
six==1.11.0
SQLAlchemy==1.2.13
stevedore==1.30.0
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
urllib3==1.23
vine==1.1.4
virtualenv==16.1.0
virtualenv-clone==0.4.0
virtualenvwrapper==4.8.2
virtualenvwrapper-win==1.2.5
visitor==0.1.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
WTForms==2.2.1
wxpy==0.3.9.8
you-get==0.4.1099
